+++ src/main/java/org/moinex/repository/CalendarEventRepository.java
@@ -11,4 +11,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CalendarEventRepository extends JpaRepository<CalendarEvent, Long> {}
+public interface CalendarEventRepository extends JpaRepository<CalendarEvent, Integer> {}
