+++ src/main/java/org/moinex/repository/CategoryRepository.java
@@ -14,7 +14,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
     /**
      * Get all non archived categories ordered by name
@@ -42,5 +42,5 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
      * @return Number of transactions
      */
     @Query("SELECT count(t) FROM WalletTransaction t WHERE t.category.id = :categoryId")
-    Long getCountTransactions(@Param("categoryId") Long categoryId);
+    Integer getCountTransactions(@Param("categoryId") Integer categoryId);
 }
