+++ src/main/java/org/moinex/repository/investment/TickerRepository.java
@@ -15,7 +15,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TickerRepository extends JpaRepository<Ticker, Long> {
+public interface TickerRepository extends JpaRepository<Ticker, Integer> {
 
     /**
      * Check if a ticker with the given name exists
@@ -37,7 +37,7 @@ public interface TickerRepository extends JpaRepository<Ticker, Long> {
      * @return The count of purchases associated with the ticker
      */
     @Query("SELECT count(p) " + "FROM TickerPurchase p " + "WHERE p.ticker.id = :tickerId")
-    Long getPurchaseCountByTicker(@Param("tickerId") Long tickerId);
+    Integer getPurchaseCountByTicker(@Param("tickerId") Integer tickerId);
 
     /**
      * Get count of sales associated with the ticker
@@ -45,7 +45,7 @@ public interface TickerRepository extends JpaRepository<Ticker, Long> {
      * @return The count of sales associated with the ticker
      */
     @Query("SELECT count(s) " + "FROM TickerSale s " + "WHERE s.ticker.id = :tickerId")
-    Long getSaleCountByTicker(@Param("tickerId") Long tickerId);
+    Integer getSaleCountByTicker(@Param("tickerId") Integer tickerId);
 
     /**
      * Get count of dividends associated with the ticker
@@ -53,7 +53,7 @@ public interface TickerRepository extends JpaRepository<Ticker, Long> {
      * @return The count of dividends associated with the ticker
      */
     @Query("SELECT count(d) " + "FROM Dividend d " + "WHERE d.ticker.id = :tickerId")
-    Long getDividendCountByTicker(@Param("tickerId") Long tickerId);
+    Integer getDividendCountByTicker(@Param("tickerId") Integer tickerId);
 
     /**
      * Get count of crypto exchanges associated with the ticker
@@ -64,7 +64,7 @@ public interface TickerRepository extends JpaRepository<Ticker, Long> {
             "SELECT count(e) "
                     + "FROM CryptoExchange e "
                     + "WHERE e.soldCrypto.id = :tickerId OR e.receivedCrypto.id = :tickerId")
-    Long getCryptoExchangeCountByTicker(@Param("tickerId") Long tickerId);
+    Integer getCryptoExchangeCountByTicker(@Param("tickerId") Integer tickerId);
 
     /**
      * Findall tickers and order them by symbol in ascending order
