+++ src/main/java/org/moinex/repository/wallettransaction/WalletRepository.java
@@ -13,7 +13,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WalletRepository extends JpaRepository<Wallet, Long> {
+public interface WalletRepository extends JpaRepository<Wallet, Integer> {
     /**
      * Check if a wallet with the given name exists
      * @param name The name of the wallet
